Seperate the logic of DiscogsFetcher to seperate class file.
